requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
matplotlib
kaggle
streamlit
Pillow